requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
motor
cloudinary
langchain
langchain-community
langchain-google-genai
langchain-cohere
PyPDFium2
faiss-cpu
